scripts/dataset/requirements.txt
feedparser
beautifulsoup4
requests
discord.py
python-dotenv
pymongo[srv]